// src/model/card-id.ts
export enum CardId
{
  PURPLE_EMPEROR = "Schillerfalter",
  LARGE_TORTOISESHELL = "Groser-Fuchs",
  GNAT = "Stechmuecke",
  BEECH = "Buche",
  CHANTERELLE = "Pfifferling",
  BULLFINCH = "Gimpel",
  PENNY_BUN = "Steinpilz",
  CAMBERWELL_BEAUTY = "Trauermantel",
  TREE_FERNS = "Baumfarn",
  CHAFFINCH = "Buchfink",
  BIRCH = "Birke",
  HEDGEHOG = "Igel",
  EUROPEAN_BADGER = "Dachs",
  GOSHAWK = "Habicht",
  POND_TURTLE = "Sumpfschildkroete",
  RED_SQUIRREL = "Eichhoernchen",
  WOLF = "Wolf",
  WILD_STRAWBERRIES = "Walderdbeeren",
  OAK = "Eiche",
  BROWN_BEAR = "Braunbaer",
  RED_DEER = "Rothirsch",
  GREAT_SPOTTED_WOODPECKER = "Buntspecht",
  SQUEAKER = "Frischling",
  FIRE_SALAMANDER = "Feuersalamander",
  FLY_AGARIC = "Fliegenpilz",
  TAWNY_OWL = "Waldkauz",
  GREATER_HORSESHOE_BAT = "Hufeisennase",
  FALLOW_DEER = "Damhirsch",
  BECHSTEIN = "Bechsteinfledermaus",
  RED_FOX = "Fuchs",
  RACCOON = "Waschbaer",
  BEECH_MARTEN = "Steinmarder",
  PEACOCK_BUTTERFLY = "Tagpfauenauge",
  WILD_BOAR = "Wildschwein",
  SILVER_FIR = "Tanne",
  SYCAMORE = "Ahorn",
  EUROPEAN_HARE = "Feldhase",
  TREE_FROG = "Laubfrosch",
  HORSE_CHESTNUT = "Kastanie",
  FIREFLIES = "Gluehwuermchen",
  BLACKBERRIES = "Brombeeren",
  DOUGLAS_FIR = "Douglasie",
  MOSS = "Moos",
  EUROPEAN_FAT_DORMOUSE = "Siebenschlaefer",
  MOLE = "Maulwurf",
  COMMON_TOAD = "Erdkroete",
  PARASOL_MUSHROOM = "Parasol",
  ROE_DEER = "Reh",
  STAG_BEETLE = "Hirschkaefer",
  LINDEN = "Linde",
  EURASIAN_JAY = "Eichelhaeher",
  BARBASTELLE_BAT = "Mopsfledermaus",
  VIOLET_CARPENTER_BEE = "Holzbiene",
  LYNX = "Luchs",
  WOOD_ANT = "Waldameise",
  BROWN_LONG_EARED_BAT = "Braunes-Langohr",
  SILVER_WASHED_FRITILLARY = "Kaisermantel",
  SAMPLING = "Baumsproessling",
  VACCINIUM_MYRTILLUS = "Heidelbeere",
  ICHTHYOSAURA_ALPESTRIS = "Bergmolch",
  CRATERELLUS_CORNUCOPIODES = "Herbsttrompete",
  LARIX = "Europaeische-Laerche",
  AQUILA_CHRYSAETOS = "Steinadler",
  PINUS = "Zirbelkiefer",
  LEPUS_TIMIDUS = "Schneehase",
  RUPICAPRA_RUPICAPRA = "Gaemse",
  GENTIANA = "Enzian",
  TETRAO_UROGALLUS = "Auerhuhn",
  PARNASSIUS_PHOEBUS = "Alpen-Apollofalter",
  MARMOTA_MARMOTA = "Alpenmurmeltier",
  CAPRA_IBEX = "Steinbock",
  CORVUS_CORAX = "Kolkrabe",
  GYPAETUS_BARBATUS = "Bartgeier",
  LEONTOPODIUM_NIVALE = "Edelweis",
  HYPSUGO_SAVII = "Alpenfledermaus",
  SQUEAKER_EDGE = "Frischling-Waldrand",
  SAMBUCUS = "Holunder",
  COMMON_HAZEL = "Haselnuss",
  BLACKTHORN = "Schlehdorn",
  WILD_BOAR_FEMALE = "Bache",
  BEEHIVE = "Bienenschwarm",
  EUROPEAN_BISON = "Wisent",
  EUROPEAN_WILDCAT = "Wildkatze",
  COMMON_PIPISTRELLE = "Zwergfledermaus",
  MOSQUITO = "Schnake",
  EUROPEAN_POLECAT = "Waldiltis",
  MAP_BUTTERFLY = "Landkaertchen",
  URTICA = "Brennnessel",
  GREAT_GREEN_BUSH_CRICKET = "Gruenes-Heupferd",
  EUROPEAN_WATER_VOLE = "Schermaus",
  EURASIAN_MAGPIE = "Elster",
  COMMON_NIGHTINGALE = "Nachtigall",
  BARN_OWL = "Schleiereule",
  DIGITALIS = "Fingerhut"
}